refactor: continuing the rip apart of custom vals
