+++ product_configurator/models/__init__.py
@@ -1,4 +1,5 @@
 from . import product_config
 from . import product_attribute
-from . import product
+from . import product_template
+from . import product_product
 from . import ir_ui_view
